Only injecting LivePage once per a page load
`chrome.tabs.onUpdated.addListener` would fire when the URL had changes
(but not necessarily the page), as a result LivePage would get injected
a bunch of times. Moving to a "content script tells extension when it's
read" approach fixed this.

Also tidied up the scripts to make testing a little easier.

Fixes #79